common: do not byte-verify a site another mod already hooked
Every mod checked its sites by comparing the bytes at each RVA against the
game's prologue. That is correct exactly once: the moment Mewjector
manages a site, those bytes are its FF 25 jump, so the check fails and the
mod refuses to install for the one reason it never should - that another
mod wanted the same function.

This bit three times before being recognised. mewlinks refused when a
family tree probe hooked HouseCatStatus::init. Migrating to MJ_InstallHook
made hooks chainable but left the check that makes chaining impossible.
Then a new mod refused because a combat probe had hooked Character::init
first.

verify_sites_chained() asks MJ_QueryHook whether a site is already managed
and, where it is, trusts the verification Mewjector did when the first
hook went in. The byte-reading version is gone rather than deprecated, so
it cannot be reached for.

Also adds no_more_overflow_flies: a write-up and fix sketch for the
Incubator+ / Dark One overflow flies, deliberately not released. A
character spawned mid-turn keeps the engine's INT_MAX/2 "full health"
sentinel in current HP because Character::init writes max but not current.
Damage landing in that window is reported as the whole sentinel-to-max
drop, so any leech heals for ~2^30 - which Incubator+ then scales a rot
fly by. 18 of 18 collapses in one fight were on same-turn spawns, none on
older ones. Filling current HP at the end of init took it to 0.

+++ mewlinks/test_abi.cpp
@@ -59,7 +59,7 @@ int main() {
     check(w.size == 0 && w.cap == 7 && w.buf[0] == 0, "wstring is empty with capacity 7");
 
     // ---- std::function ------------------------------------------------------
-    void* const comp = (void*)0xDEADBEEF;
+    void* const comp = (void*)(size_t)0xDEADBEEF;   // size_t: MSVC warns C4312 otherwise
     MsvcFunction fn;
     fn_init(&fn, comp, 1, 0);
     check((void*)fn.impl == (void*)fn.storage, "impl is stored inline, as the game's lambdas are");
